Issue 23909: add export format to initialization of email templates
