+++ api-src/org/labkey/api/workflow/DataManagerConfig.java
@@ -11,15 +11,14 @@ import java.util.Map;
  * Created by davebradlee on 7/20/15.
  *
  */
-public abstract class DataManagerConfig
+public abstract class DataManagerConfig extends WorkflowDelegateActionBase
 {
-    protected final Map<String, Object> _variables;
     private final Container _container;
     private final User _initiator;
 
     public DataManagerConfig(Map<String, Object> variables)
     {
-        _variables = variables;
+        super(variables);
         _container = ContainerManager.getForId((String) _variables.get(WorkflowProcess.CONTAINER_ID));
         _initiator = UserManager.getUser(Integer.valueOf((String) _variables.get(WorkflowProcess.INITIATOR_ID)));
     }
@@ -36,43 +35,4 @@ public abstract class DataManagerConfig
         return _initiator;
     }
 
-    public Object getVariable(String key)
-    {
-        return _variables.get(key);
-    }
-
-    public void setVariable(String key, Object value)
-    {
-        _variables.put(key, value);
-    }
-
-    public Map<String, Object> getVariables()
-    {
-        return _variables;
-    }
-
-    @SuppressWarnings("unchecked")
-    public Object getDataAccessParameter(String key)
-    {
-        Object dataAccess = _variables.get(WorkflowProcess.DATA_ACCESS_KEY);
-        if (null != dataAccess && dataAccess instanceof Map)
-        {
-            Object parameters = ((Map<String, Object>) dataAccess).get(WorkflowProcess.DATA_ACCESS_PARAMETERS_KEY);
-            if (null != parameters && parameters instanceof Map)
-                return ((Map<String, Object>)parameters).get(key);
-        }
-        return null;
-    }
-
-    @SuppressWarnings("unchecked")
-    public void setDataAccessParameter(String key, Object value)
-    {
-        Object dataAccess = _variables.get(WorkflowProcess.DATA_ACCESS_KEY);
-        if (null != dataAccess && dataAccess instanceof Map)
-        {
-            Object parameters = ((Map<String, Object>) dataAccess).get(WorkflowProcess.DATA_ACCESS_PARAMETERS_KEY);
-            if (null != parameters && parameters instanceof Map)
-                ((Map<String, Object>)parameters).put(key, value);
-        }
-    }
 }
